requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
gunicorn==23.0.0